==> notebooks/1.0-mg-EDA.ipynb <==
from collections import defaultdict
from sklearn.model_selection import train_test_split
from typing import List

import pandas as pd

data = pd.read_csv('../data/raw/train.csv')
data.head()

## Соотношение классов

data['is_duplicate'].value_counts()

Видим очень сильный дисбаланс в сторону класса 0 (99.3%)

### Класс 1

data[data['is_duplicate']==1]

### Сколько всего уникальных названий компаний

unique_companies = pd.concat([data['name_1'], data['name_2']]).unique()
len(unique_companies)

### Сколько из них встречаюся в классе 1:

name_1_duplicates = data.loc[data['is_duplicate'] == 1, 'name_1']
name_2_duplicates = data.loc[data['is_duplicate'] == 1, 'name_2']
unique_companies_class_1 = pd.concat([name_1_duplicates, name_2_duplicates]).unique()
len(unique_companies_class_1)

## Анализ на уровне символов

symbols = set()
for name in unique_companies:
    symbols.update(name.lower())

symbols

Помимо англ и русских букв встречается много знаков препинания, цифры + китайский и арабский(?) алфавит. Посмотрим насколько это частое явление.

comps = set()
for s in ['\u0e00', '\u3000']:
    comps.update([name for name in unique_companies if s in name])
comps

# Арабский символы

companies_with_arabic_symbols = set()
for s in ['ر', 'س', 'ف', 'ك', 'م', 'و', 'ي']:
    companies_with_arabic_symbols.update([name for name in unique_companies if s in name])
companies_with_arabic_symbols

Всего 1 компания с арабскими символами в названии

# Китайские иероглифы

companies_with_chinese_symbols = set()
for s in ['上', '东', '京', '份', '会', '公', '北', '双', '司', '团', '式', '彤', '技', '料', '新', '方', '日', '有', '术', '材', '株', '水', '海', '社', '程', '股', '虹', '防', '限', '集', '雨']:
    companies_with_chinese_symbols.update([name for name in unique_companies if s in name])
companies_with_chinese_symbols

Всего 4 компании с китайскими иероглифами в названии

# Турецкие символы (?)

companies_with_turkish_symbols = set()
for s in ['á',
 'ã',
 'ç',
 'è',
 'é',
 'í',
 'ñ',
 'ó',
 'õ',
 'ö',
 'ú',
 'ü',
 'ę',
 'ł',
 'ő',
 'ş',
 'ű',
 '̇']:
    companies_with_turkish_symbols.update([name for name in unique_companies if s in name])
companies_with_turkish_symbols

Пока оставим как есть

companies_with_numbers = set()
for s in ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9']:
    companies_with_numbers.update([name for name in unique_companies if s in name])
companies_with_numbers

Кажется, что цифры в названиях чаще являются шумом (возможно их стоит тоже почистить)

## Класс 1

class_1 = data[data['is_duplicate']==1].reset_index(drop=True)
class_1

## Оказывается есть и такие случаи

class_1[class_1['name_1'] == class_1['name_2']]

class_1 = class_1[class_1['name_1'] != class_1['name_2']].reset_index(drop=True)

name_1_duplicates = class_1['name_1']
name_2_duplicates = class_1['name_2']
unique_companies_class_1 = pd.concat([name_1_duplicates, name_2_duplicates]).unique()
len(unique_companies_class_1)
