==> notebooks/1.1-pk-feature-extraction.ipynb <==
# Вычленение фичей на основании схожести строк

import re
import math
from collections import Counter

import pandas as pd
from fuzzywuzzy import fuzz
import Levenshtein

data = pd.read_csv(
    '../data/processed/unificated_train.csv', 
    index_col='pair_id'
)

data.sample(5)

## Вычленение фичей из транслитерированных названий

# Ratio of similar characters on both string
data['trans_wratio'] = data.apply(
    lambda row: fuzz.WRatio(
        row.transliterated_name_1,
        row.transliterated_name_2
    ) / 100, 
    axis = 1
)

# Ratio of the most similar substring
data['trans_partial_ratio'] = data.apply(
    lambda row: fuzz.partial_ratio(
        row.transliterated_name_1,
        row.transliterated_name_2
    ) / 100, 
    axis = 1
)

# Measure of the sequences' tokens similarity
data['trans_token_sort_ratio'] = data.apply(
    lambda row: fuzz.token_sort_ratio(
        row.transliterated_name_1,
        row.transliterated_name_2
    ) / 100,
    axis = 1
)

# Levenshtein distance
data['trans_levenshtein'] = data.apply(
    lambda row: Levenshtein.distance(
        row.transliterated_name_1,
        row.transliterated_name_2
    ),
    axis = 1
)

# Levenshtein distance normalized to the maximum length
data['trans_levenshtein_ratio'] = data.apply(
    lambda row: Levenshtein.distance(
        row.transliterated_name_1,
        row.transliterated_name_2
    ) / max(
        len(row.transliterated_name_1),
        len(row.transliterated_name_2)
    ), 
    axis = 1
)

# Jaro distance
data['trans_jaro'] = data.apply(
    lambda row: Levenshtein.jaro(
        row.transliterated_name_1,
        row.transliterated_name_2
    ),
    axis = 1
)

def text_to_vector(text: str) -> Counter:
    '''
    Transform text to vector
    
    Args:
        text: Original text
        
    Returns:
        Corresponding vector
    '''
    word = re.compile(r'\w+')
    words = word.findall(text)
    
    return Counter(words)


def get_cos_distance(string_1: str, string_2: str) -> float:
    '''
    Get cosine distance between two strings
    
    Args:
            string_1: First string
            string_2: Second string
            
    Returns:
        Cosine distance between vectors
    '''
    def _get_cosine(vec_1: Counter, vec_2: Counter) -> float:
        '''
        Get cosine distance between two vectors
        
        Args:
            vec_1: First vector
            vec_2: Second vector
            
        Returns:
            Cosine distance between vectors
        '''
        intersection = set(vec_1.keys()) & set(vec_2.keys())
        numerator = sum([vec_1[x] * vec_2[x] for x in intersection])

        sum1 = sum([vec_1[x]**2 for x in vec_2.keys()])
        sum2 = sum([vec_1[x]**2 for x in vec_2.keys()])
        
        denominator = math.sqrt(sum1) * math.sqrt(sum2)

        if not denominator:
            return 0.0
        else:
            return float(numerator) / denominator
        
    vector_1 = text_to_vector(string_1)
    vector_2 = text_to_vector(string_2)

    return _get_cosine(vector_1, vector_2)

# Cosine distance between strings
data['trans_cosine'] = data.apply(
    lambda row: get_cos_distance(
        row.transliterated_name_1, 
        row.transliterated_name_2
    ), 
    axis = 1
)

def jaccard_similarity(string_1: str, string_2: str) -> float:
    '''
    Get jaccard similarity of two strings

    Args:
      string_1: First string
      string_2: Second string

    Returns:
        Jaccard similarity of string
    '''
    intersection_cardinality = len(set.intersection(*[set(string_1), set(string_2)]))
    union_cardinality = len(set.union(*[set(string_1), set(string_2)]))
    
    return intersection_cardinality/float(union_cardinality)

# Jaccard similarity of strings
data['trans_jaccard'] = data.apply(
    lambda row: jaccard_similarity(
        row.transliterated_name_1, 
        row.transliterated_name_2
    ), 
    axis = 1
)

## Вычленение фичей из очищенных и конкатенированных названий

# Ratio of similar characters on both string
data['conc_wratio'] = data.apply(
    lambda row: fuzz.WRatio(
        row.concated_name_1,
        row.concated_name_2
    ) / 100, 
    axis = 1
)

# Ratio of the most similar substring
data['conc_partial_ratio'] = data.apply(
    lambda row: fuzz.partial_ratio(
        row.concated_name_1, 
        row.concated_name_2
    ) / 100,
    axis = 1
)

# Measure of the sequences' tokens similarity
data['conc_token_sort_ratio'] = data.apply(
    lambda row: fuzz.token_sort_ratio(
        row.concated_name_1, 
        row.concated_name_2
    ) / 100,
    axis = 1
)

# Levenshtein distance
data['conc_levenshtein'] = data.apply(
    lambda row: Levenshtein.distance(
        row.concated_name_1, 
        row.concated_name_2
    ), 
    axis = 1
)

# Levenshtein distance normalized to the maximum length
data['conc_levenshtein_ratio'] = data.apply(
    lambda row: Levenshtein.distance(
        row.concated_name_1, 
        row.concated_name_2
    ) / max(
        len(row.concated_name_1),
        len(row.concated_name_2)
    ), 
    axis = 1
)

# Jaro distance
data['conc_jaro'] = data.apply(
    lambda row: Levenshtein.jaro(
        row.transliterated_name_1, 
        row.concated_name_2
    ), 
    axis = 1
)

## Сохранение датафрейма с фичами

# Dropping all non-numerical columns
data_numerical = data.drop(
    ['name_1',
     'name_2',
     'concated_name_1',
     'concated_name_2',
     'transliterated_name_1',
     'transliterated_name_2'
    ],
    axis=1
)

data_numerical.head()

# Saving of preprocessed dataset
data_numerical.to_csv('../data/processed/numerical_train.csv')
