==> notebooks/2.1-mg-stop-words.ipynb <==
# Добавляем путь к основной папке проекта, чтобы иметь возможность делать импорт из src
import os
import sys
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)

from collections import Counter

import pandas as pd
from src.data.rnn_preprocessing import clean_company_name_string

data = pd.read_csv('../data/raw/train.csv')

names = list(data['name_1'].values) + list(data['name_2'])
names = sorted(set(names))

clean_names = [clean_company_name_string(n) for n in names]
print(f'Всего названий {len(clean_names)}')

words = []
for n in clean_names:
    words.extend(n.split())

stop_words = [word for word, cnt in Counter(words).most_common() if cnt > 70]
len(stop_words)

stop_words

Counter(words).most_common()
